Initial Release
It should be noted that this is majorly alpha, and not very functional.
There is no option for post messages, no user registration system, and
SQL queries are pretty inefficient etc. So yeah - not that you could
but you shouldn't use this in a production environment without major
modification.

It's a practice run for a hackday. After the hack day, a much more
polished product should be released.