Need account to view posts
Until we've figured a way to control which posts can and cannot be
viewed (implement "public post" feature), only users now need accounts
to view posts, and can only view posts that match the institution they
chose when registering their account.

In addition the above, there is now an option to go "Back to listings"
when viewing an individual post.